Move some file needed by codegen in Abstractions and change Orleans usage detection in codegen
