+++ src/Orleans.Core.Abstractions/SystemTargetInterfaces/ISystemTargetBase.cs
@@ -9,6 +9,6 @@ namespace Orleans
     {
         SiloAddress Silo { get; }
         GrainId GrainId { get; }
-        IRuntimeClient RuntimeClient { get; }
+        IGrainReferenceRuntime GrainReferenceRuntime { get; }
     }
 }
